Rename TokenizedBuffer to TextMateLanguageMode
